+++ tests/gaggle.rs
@@ -10,11 +10,11 @@ const INDEX_PATH: &str = "/";
 const ABOUT_PATH: &str = "/about.html";
 
 pub async fn get_index(user: &GooseUser) {
-    let _response = user.get(INDEX_PATH).await;
+    let _goose = user.get(INDEX_PATH).await;
 }
 
 pub async fn get_about(user: &GooseUser) {
-    let _response = user.get(ABOUT_PATH).await;
+    let _goose = user.get(ABOUT_PATH).await;
 }
 
 /// Test test_start alone.