+++ tests/logs.rs
@@ -12,19 +12,19 @@ const STATS_LOG_FILE: &str = "stats.log";
 const DEBUG_LOG_FILE: &str = "debug.log";
 
 pub async fn get_index(user: &GooseUser) {
-    let _response = user.get(INDEX_PATH).await;
+    let _goose = user.get(INDEX_PATH).await;
 }
 
 pub async fn get_error(user: &GooseUser) {
-    let response = user.get(ERROR_PATH).await;
-    if let Ok(r) = response.response {
+    let goose = user.get(ERROR_PATH).await;
+    if let Ok(r) = goose.response {
         let headers = &r.headers().clone();
         match r.text().await {
             Ok(_) => {}
             Err(_) => {
                 user.log_debug(
                     "there was an error",
-                    Some(response.request),
+                    Some(goose.request),
                     Some(headers),
                     None,
                 );